requirements.txt
paramiko==3.5
python-dotenv